+++ packages/filterlist/src/eventTypes.ts
@@ -16,7 +16,7 @@ export const resetAllFilters: EventType = "resetAllFilters";
 export const reload: EventType = "reload";
 export const setSorting: EventType = "setSorting";
 export const resetSorting: EventType = "resetSorting";
-export const setFiltersAndSorting: EventType = "setFiltersAndSorting";
+export const updateStateAndRequest: EventType = "updateStateAndRequest";
 
 export const changeLoadParams: EventType = "changeLoadParams";